Add sv_lanonly and server_list fix
sv_lanonly 1 will not send heartbeat to master.

server_list: when you closed dedicated with net_port 27016 and then right after opened a client with the same net_port, the server list will query the client with getInfo and then you would see that on the server list. now it only shows if the server is dedicated too.