Fix service not available disconnect
When loading the map takes too long the game can give the error:
"The Call of Duty: Advanced Warfare service is not available at this time. Please try again later."

This happens often on lower end computers.
Live_CheckForFullDisconnect seems to be the cause for this, which I have patched for this fix.